src-command/consumer/requirements.txt
kafka-python==2.3.0
pymongo==4.16.0
prometheus-client==0.23.1
